pyproject.toml
[tool.poetry]
name = "shortify"
version = "0.1.0"
description = ""
authors = ["oskar <oskarkorgul@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]


[[tool.poetry.source]]
name = "fastapi-cbv"
url = "https://gitlab.com/api/v4/projects/43600871/packages/pypi/simple"
default = false
secondary = true

[tool.poetry.dependencies]
python = "^3.10"
fastapi-cbv = "^0.4"
hypercorn = "^0.14.3"
aioredis = "^2.0.1"
nanoid = "^2.0.0"
motor = "^3.1.2"
orjson = "^3.8.10"
aiokafka = "^0.8.0"
aiorun = "^2022.11.1"
clickhouse-connect = "^0.5.23"
user-agents = "^2.2.0"
geoip2 = "^4.6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
fakeredis = "^2.10.3"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"
mongomock-motor = "^0.0.18"
faker = "^18.4.0"
coverage = "^7.2.3"
pytest-timeout = "^2.1.0"

[tool.pytest.ini_options]
asyncio_mode= "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
